example: Add web color register tutorial
The web color tutorial aims to walk a user through the basic commands.
From exploring the current state of a register to adding data and
finally publishing it to Netlify.

The netlify configuration lacks proper content negotiation and it is
likely that it is not possible. It is still a good tradeoff.

Signed-off-by: Arnau Siches <arnau.siches@digital.cabinet-office.gov.uk>