feat: Seal RSF patches
A sealed patch is a patch where the first command is a root hash
assertion for the expected root hash in the register and the last one is
the final root hash after applying the patch to the register.

Sealed patches are expected to reinforce guarantees where a patch is
applied to the expected register state.

Sealed patches are only for data as metadata is not part of the log
assertion.

Signed-off-by: Arnau Siches <arnau.siches@digital.cabinet-office.gov.uk>