Set some sane defaults for CF Manifest
The default memory is resolving to 1GB. This, although optimistic, is
probably not necessary. With an idea of serving static files, we can
save some resources to the GOV.UK PaaS team.

The number of instances for production, is recommended to be 2+. The
default is sadly 1. The reason behind this, is that the apps can be
taken away and spun elsewhere due to some maintenance or issues. Having
more, means potentially having less downtime.